requirements.txt
python-telegram-bot>=21.0,<22.0
yt-dlp>=2025.1.0
